Fix format_percent to handle string floats and add unit tests (#586)
- Updated format_percent in gptscan.py to use float() instead of int().
- Added comprehensive unit tests for format_percent in tests/test_config_and_utils.py.
- Verified that 85.5 and "85.5" are correctly formatted as "86%".
- Ensured all tests pass in the required Python 3.10.20 environment.

+++ gptscan.py
@@ -1738,7 +1738,7 @@ def parse_percent(val: str, default: float = -1.0) -> float:
 def format_percent(val: Any) -> str:
     """Format a numeric threat level (0-100) as a percentage string (e.g., 85 to "85%")."""
     try:
-        return "{:.0%}".format(int(val) / 100.)
+        return "{:.0%}".format(float(val) / 100.)
     except (ValueError, TypeError):
         return "Error"
